ctrl+click opens in a new tab bug fixed
