+++ src/pages/Explorer/Layouts/LayoutItem.svelte
@@ -1,7 +1,8 @@
 <script>
   import SidebarItem from '../Layouts/SidebarItem.svelte'
   import ExplorerItem from '../Layouts/ExplorerItem.svelte'
-  import { getItemUrl } from '../const'
+  import { getItemRoute, getItemUrl } from '../const'
+  import { history } from '../../../redux'
 
   export let small = false
   export let item
@@ -9,10 +10,33 @@
   export let showActions = false
   export let hasIcons = false
   export let assets = []
+
+  const ACTION_BUTTON_CLASS = 'actionbutton'
+
+  $: url = getItemUrl(item, type)
+
+  function onClick(e) {
+    const isIcon = e.target && ['use', 'svg'].includes(e.target.tagName)
+
+    const isActionButton =
+      e.target.classList.contains(ACTION_BUTTON_CLASS) ||
+      (e.target.parentElement && e.target.parentElement.classList.contains(ACTION_BUTTON_CLASS))
+
+    if (isIcon || isActionButton) {
+      e.preventDefault()
+      return
+    }
+
+    if (!e.ctrlKey && url.includes(location.hostname)) {
+      e.preventDefault()
+      history.push(getItemRoute(item, type))
+      return
+    }
+  }
 </script>
 
 {#if small}
-  <SidebarItem {item} {small} {type} {showActions} url={getItemUrl(item, type)} />
+  <SidebarItem {item} {small} {showActions} {url} {onClick} />
 {:else}
-  <ExplorerItem {item} {type} url={getItemUrl(item, type)} {hasIcons} {assets} />
+  <ExplorerItem {item} {type} {url} {hasIcons} {assets} {onClick} />
 {/if}
